# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0"
    },
    "paths": {
        "/api/user": {
            "get": {
                "operationId": "961ad57987905b44ae2e2f40b82e76de",
                "responses": {
                    "200": {
                        "description": "Get current user"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "John Doe",
                                    "email": "john@example.com",
                                    "password": "password123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register a new user"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john@example.com",
                                    "password": "password123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Log in a user"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Get a list of products"
                    }
                }
            },
            "post": {
                "operationId": "af9ea3f02fbf732a37e94ab4cf3f94c7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Product Name",
                                    "description": "Product Description"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new product"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get a specific product"
                    }
                }
            },
            "put": {
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Updated Product Name",
                                    "description": "Updated Product Description"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a specific product"
                    }
                }
            },
            "delete": {
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete a specific product"
                    }
                }
            }
        }
    }
}